+++ CHANGES.md
@@ -16,10 +16,19 @@
 - `CameraCapturerFactory` にフロント/リアカメラの優先順位のオプションを追加した
 - サイマルキャスト配信のエンコーダ設定変更用コールバックを追加した
   - `SoraMediaChannel.Listener#onSenderEncodings()`
+- 定数 `SoraErrorReason.ICE_DISCONNECTED` を追加した
+- `SoraMediaChannel.Listener` に `onWarning` メソッドを追加した
+  - このバージョンでは `ICE_DISCONNECTED` の通知のみに利用している
+  - 想定ユースケースは、ネットワークが不安定であることを UI に伝えること
+  - デフォルト実装は処理なしである
 
 ### UPDATE
 
 - `com.android.tools.build:gradle` を 3.5.3 に上げた
+
+### FIX
+
+- IceConnectionState = disconnected では切断処理を行わないよう変更した
 
 ## 1.10.0